FE/Qt: Medium related stuff: Moving StorageSlot from UIDefs to UIMediumDefs to avoid unwanted dependencies; This allows UIMachineDefs as well to detach from using UIDefs.
svn:sync-xref-src-repo-rev: r161648

+++ src/VBox/Frontends/VirtualBox/src/settings/editors/UIStorageSettingsEditor.h
@@ -1,4 +1,4 @@
-/* $Id: UIStorageSettingsEditor.h 103338 2024-02-13 16:37:06Z sergey.dubov@oracle.com $ */
+/* $Id: UIStorageSettingsEditor.h 103339 2024-02-13 16:49:47Z sergey.dubov@oracle.com $ */
 /** @file
  * VBox Qt GUI - UIStorageSettingsEditor class declaration.
  */
@@ -104,9 +104,9 @@ struct UIDataStorageAttachment
     /** Holds the device type. */
     KDeviceType  m_enmDeviceType;
     /** Holds the port. */
-    LONG         m_iPort;
+    qint32       m_iPort;
     /** Holds the device. */
-    LONG         m_iDevice;
+    qint32       m_iDevice;
     /** Holds the medium ID. */
     QUuid        m_uMediumId;
     /** Holds whether the attachment being passed through. */
